Registered the Books model to the books app admin
